Move Writer selection offset helpers to text_helpers.py
Move _writer_char_count, _char_offset_of_position,
_get_writer_selection_positions, and get_selection_range from
document_helpers.py to text_helpers.py so LibrePy export code paths
do not import document_helpers. Update document_helpers.py to call
the moved functions via _text_helpers alias. Update format.py,
html_export.py, AGENTS.md, and relevant test cases.
