+++ tests/writer/test_document_context_for_chat.py
@@ -41,7 +41,7 @@ def test_inject_markers_no_overlap():
 
 
 @patch("plugin.doc.document_helpers._read_writer_text_slice")
-@patch("plugin.doc.document_helpers._writer_char_count", return_value=20000)
+@patch("plugin.doc.text_helpers._writer_char_count", return_value=20000)
 @patch("plugin.doc.doc_type.get_document_type")
 def test_get_document_context_reads_slices_not_full_doc(mock_doc_type, mock_char_count, mock_read_slice):
     from plugin.doc.doc_type import DocumentType
@@ -57,7 +57,7 @@ def test_get_document_context_reads_slices_not_full_doc(mock_doc_type, mock_char
 
 
 @patch("plugin.doc.document_helpers._read_writer_text_slice")
-@patch("plugin.doc.document_helpers._writer_char_count", return_value=100)
+@patch("plugin.doc.text_helpers._writer_char_count", return_value=100)
 @patch("plugin.doc.doc_type.get_document_type")
 def test_get_document_context_short_doc_single_slice(mock_doc_type, mock_char_count, mock_read_slice):
     from plugin.doc.doc_type import DocumentType
